pyproject.toml
[tool.poetry]
name = "it_factory_test"
version = "0.1.0"
description = ""
authors = ["vildan-valeev <vildan.maydanovich.valeev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
psycopg2-binary = "^2.9.3"
python-decouple = "^3.6"
django-stubs-ext = "^0.4.0"
django-split-settings = "^1.1.0"
flake8 = "^4.0.1"
drf-yasg = "^1.20.0"
django-cors-headers = "^3.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
